Now that it all compiles with hipcc, move back to headers only
