+++ libpigi/constants.h
@@ -1,3 +1,5 @@
+#pragma once
+
 namespace Constants {
     const int c {299792458};
 }
