--- libpigi/degridop.h
@@ -1,3 +1,0 @@
-#pragma once
-
-enum class DegridOp {Replace, Subtract, Add};
