+++ libpigi/fits.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <array>
 #include <string_view>
